// src/types/categories/insect.ts
export type Insect = {
  id?: number;
  name?: string;
  iconimage?: string;
  critterpediaimage?: string;
  furnitureimage?: string;
  size?: string;
  sell?: number;
  wherehow?: string;
  weather?: string;
  totalcatchestounlock?: number;
  spawnrates?: string;
  nhjan?: string;
  nhfeb?: string;
  nhmar?: string;
  nhapr?: string;
  nhmay?: string;
  nhjun?: string;
  nhjul?: string;
  nhaug?: string;
  nhsep?: string;
  nhoct?: string;
  nhnov?: string;
  nhdec?: string;
  shjan?: string;
  shfeb?: string;
  shmar?: string;
  shapr?: string;
  shmay?: string;
  shjun?: string;
  shjul?: string;
  shaug?: string;
  shsep?: string;
  shoct?: string;
  shnov?: string;
  shdec?: string;
  color1?: string;
  color2?: string;
  iconfilename?: string;
  critterpediafilename?: string;
  furniturefilename?: string;
  internalid?: number;
  uniqueentryid?: string;
};